# Dive pictures: Don't plot pictures twice when changing current dive