--- a/profile-widget/profilewidget2.h
+++ b/profile-widget/profilewidget2.h
@@ -73,7 +73,7 @@ public:
 
 	ProfileWidget2(QWidget *parent = 0);
 	void resetZoom();
-	void plotDive(struct dive *d = 0, bool force = false);
+	void plotDive(struct dive *d = 0, bool force = false, bool clearPictures = false);
 	void setupItem(AbstractProfilePolygonItem *item, DiveCartesianAxis *vAxis, int vData, int hData, int zValue);
 	void setPrintMode(bool mode, bool grayscale = false);
 	bool getPrintMode();