In MainWindow::current_dive_changed() first plotDive() is called,
which replots all the pictures by calling plotPictures(). This
is pointess, because it plots the pictures of the previous dive.

Then, updateDiveInfo() is called, which resets the dive pictures
and automatically replots them. Thus, switching between dives
both with hundreds of pictures is way slower than necessary.

Switching the plotDive() and updateDiveInfo() calls doesn't work.
The reason is not 100% clear, but it doesn't make sense to plot
pictures of the new dive as long as the profile still shows the
old dive anyway.

As a quick-fix, add a flag to plotDive(), which tells the function
to clear the pictures list instead of redrawing it.
Ultimately, plotDive() should probably be split in two functions.
One for the callers who update the pictures themselves and one
for the others.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>

<!-- Lines like this one are comments and will not be shown in the final output. -->
<!-- Make sure that you have read the "Contributing" section of the README and also the notes in CodingStyle. -->
<!-- If you are a collaborator, please add labels and assign other collaborators for a review. -->

### Describe the pull request:
<!-- Replace [ ] with [x] to select options. -->
- [ ] Bug fix
- [ ] Functional change
- [ ] New feature
- [x] Code cleanup
- [ ] Build system change
- [ ] Documentation change
- [ ] Language translation

### Pull request long description:
<!-- Describe your pull request in detail. -->
This is an optimization of the profile plot code: On dive-change pictures of the *old* dive were plot first, which seems quite pointless.

The fix is quite disgusting in that it introduces an additional bool-parameter to `plotDives()`, which controls its behavior. But the whole `DivePictureModel` <-> `ProfileWidget2` interface is horribly messy to begin with. I don't dare to make in-depth changes to this.

I some quick benchmarking for the `plotDive()`/`updateDiveInfo()` pair of calls. I switched continuously between two dives:

Before patch
```
dive changed to: 15
Elapsed: 135 msec
dive changed to: 14
Elapsed: 142 msec
dive changed to: 15
Elapsed: 133 msec
dive changed to: 14
Elapsed: 147 msec
dive changed to: 15
Elapsed: 132 msec
dive changed to: 14
Elapsed: 139 msec
dive changed to: 15
Elapsed: 134 msec
dive changed to: 14
Elapsed: 147 msec
dive changed to: 15
Elapsed: 137 msec
dive changed to: 14
Elapsed: 142 msec
```

After patch
```
dive changed to: 15
Elapsed: 63 msec
dive changed to: 14
Elapsed: 135 msec
dive changed to: 15
Elapsed: 64 msec
dive changed to: 14
Elapsed: 134 msec
dive changed to: 15
Elapsed: 65 msec
dive changed to: 14
Elapsed: 135 msec
dive changed to: 15
Elapsed: 64 msec
dive changed to: 14
Elapsed: 135 msec
dive changed to: 15
Elapsed: 63 msec
dive changed to: 14
Elapsed: 136 msec
dive changed to: 15
Elapsed: 64 msec
dive changed to: 14
Elapsed: 135 msec
```

Note that the dive with internal id 14 has a bunch of pictures on the profile plot, where as id 15 only a few. Therefore you see a huge improvement when switching 14->15, but not for 15->14.

Looks like the data is quite convincing. :)
### Changes made:
<!-- Enumerate the changes with 1), 2), 3) etc. -->
<!-- Ensure the test cases are updated if needed. -->

### Related issues:
<!-- Reference issues with #<issue-num>. -->
<!-- Write "Fixes #<issue-num" to notify Github that this PR fixes an issue. -->

### Additional information:
<!-- Include sample dive log or other relevant information to allow testing the change where feasible. -->

### Release note:
<!-- Describe if this change needs a release note present in CHANGELOG.md. -->
<!-- Also, please make sure to add the release note on top of the file CHANGELOG.md. -->

### Documentation change:
<!-- If this PR makes changes to user functionality, then the documentation has to be updated too. -->
<!-- Please, briefly outline here what has changed in terms of the user experience (UX). -->
<!-- If UX changes have been made, a maintainer should apply the 'needs-documentation-change' label. -->

### Mentions:
<!-- Mention users that you want to review your pull request with @<user-name>. Leave empty if not sure. -->